requirements.txt
evadb
python-dotenv
pylint
# eva-decord